תיקון שגיאה בסידור ההגדרות
