+++ lib/settings/dialogs/library_settings_dialog.dart
