+++ lib/settings/panels/editor_settings_panel.dart
