+++ lib/settings/tabs/settings_tabs.dart
@@ -2,9 +2,9 @@
 export 'design_settings_tab.dart';
 export 'reading_settings_tab.dart';
 export 'library_settings_tab.dart';
-export 'package:otzaria/tools/calendar/calendar_settings_tab.dart';
-export 'gematria_settings_tab.dart';
-export 'editor_settings_tab.dart';
+export 'package:otzaria/settings/panels/calendar_settings_panel.dart';
+export '../panels/gematria_settings_panel.dart';
+export '../panels/editor_settings_panel.dart';
 export 'tools_settings_tab.dart';
 export 'package:otzaria/shortcuts/shortcuts_settings_tab.dart';
 export 'system_settings_tab.dart';
