+++ lib/settings/tabs/tools_settings_tab.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:otzaria/tools/calendar/ulits/calendar_cubit.dart';
-import 'package:otzaria/tools/calendar/calendar_settings_tab.dart';
-import 'package:otzaria/settings/tabs/gematria_settings_tab.dart';
-import 'package:otzaria/settings/tabs/editor_settings_tab.dart';
+import 'package:otzaria/settings/panels/calendar_settings_panel.dart';
+import 'package:otzaria/settings/panels/gematria_settings_panel.dart';
+import 'package:otzaria/settings/panels/editor_settings_panel.dart';
 
 /// טאב כלים — לוח שנה, גימטריות, עורך.
 ///
